pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentwork"
version = "0.1.0"
description = "Design-time evaluation framework scaffold for AI agents"
readme = "agent-work-os-prd-short.md"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
full = [
  "fastapi>=0.115.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.0",
  "typer>=0.12.0",
  "uvicorn>=0.30.0",
]

[project.scripts]
agentwork = "agentwork.cli.main:main"

[tool.setuptools]
include-package-data = true

